+++ src/test/java/com/samcomo/dbz/report/service/ReportServiceTest.java
@@ -6,7 +6,7 @@ import com.samcomo.dbz.global.s3.constants.ImageUploadState;
 import com.samcomo.dbz.global.s3.service.S3Service;
 import com.samcomo.dbz.member.model.entity.Member;
 import com.samcomo.dbz.member.model.repository.MemberRepository;
-import com.samcomo.dbz.notification.service.NotificationServiceImpl;
+import com.samcomo.dbz.notification.service.impl.NotificationServiceImpl;
 import com.samcomo.dbz.report.exception.ReportException;
 import com.samcomo.dbz.report.model.constants.PetType;
 import com.samcomo.dbz.report.model.constants.ReportStatus;
